requirements.txt
gcsa
beautiful-date
simplegmail
python-dotenv
groq